backend/requirements.txt
fastapi
uvicorn
neo4j
python-jose
passlib[bcrypt]
python-dotenv
pydantic[email]
pandas
neontology
llama-index-embeddings-azure-openai
llama-index-llms-azure-openai
llama-index-llms-bedrock
llama-index-embeddings-bedrock
llama-index-readers-file
qdrant-client
llama-index-vector-stores-qdrant
python-multipart
openpyxl
neo4j-graphrag
playwright